flask_blog/requirements.txt
alembic==1.0.11
amqp==2.5.1
aniso8601==7.0.0
anyjson==0.3.3
bcrypt==3.1.7
billiard==3.6.1.0
blinker==1.4
celery==4.3.0
celery-with-redis==3.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cssmin==0.2.0
defusedxml==0.6.0
dnspython==1.16.0
dominate==2.3.5
eventlet==0.25.1
Flask==1.1.1
Flask-Admin==1.5.3
Flask-Assets==0.12
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Cache==0.13.1
Flask-Celery-Helper==1.1.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-OAuth==0.12
Flask-OAuthlib==0.9.5
Flask-Principal==0.4.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
greenlet==0.4.15
HTMLParser==0.0.2
httplib2==0.13.1
idna==2.8
importlib-metadata==0.19
itsdangerous==1.1.0
Jinja2==2.10.1
jsmin==2.2.2
kombu==4.6.4
Mako==1.0.13
MarkupSafe==1.1.1
monotonic==1.5
oauth2==1.9.0.post1
oauthlib==2.1.0
Pillow==6.1.0
pycparser==2.19
PyMySQL==0.9.3
python-dateutil==2.8.0
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.2
redis==3.2.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
SQLAlchemy==1.3.5
urllib3==1.25.3
vine==1.3.0
visitor==0.1.3
webassets==0.12.1
Werkzeug==0.15.4
WTForms==2.2.1
zipp==0.5.2
